Use github actions to publish images
Dockerhub has an important latency to build docker images, we have seen
up to 400 minutes, and does not provide significant tooling to re-build
only images that have changed.

In this commit, we get to build only images that were changed, either in
the Pull Request or in the pushed branch.

Running al builds in parallel has been taking, in github actions around
10 minutes.

All image builds are requested in parallel, and thanks to the dockerhub
integration that automatically creates image repositories, any new image
will automatically be pushed, fixing #7

Fixes: #7